requirements.txt
beautifulsoup4==4.12.3
EbookLib==0.18
mutagen==1.47.0
openai==1.35.7
requests==2.32.3
socksio==1.0.0
edge-tts==6.1.12
pydub==0.25.1
python-dotenv~=1.0.1
langdetect~=1.0.9
pandas~=2.2.2
python-telegram-bot~=21.3